requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
groq
python-dotenv==1.0.1
streamlit
